api/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
APScheduler==3.10.4
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.3
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
coverage==5.5
distlib==0.3.7
Faker==14.0.0
filelock==3.12.2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT==0.2.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==1.1.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.8
Werkzeug==3.0.0
WTForms==3.0.1
